--- app_cs/app/app.http
@@ -1,6 +1,0 @@
-@App_HostAddress = http://localhost:5184
-
-GET {{App_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
