--- app_cs/app_cs.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App", "App\App.csproj", "{272A4F23-A068-4BAD-AC21-523A10A335AC}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.Tests", "App.Tests\App.Tests.csproj", "{92A7A7C4-CCDC-417E-84BA-B06F485630C5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{272A4F23-A068-4BAD-AC21-523A10A335AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{272A4F23-A068-4BAD-AC21-523A10A335AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{272A4F23-A068-4BAD-AC21-523A10A335AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{272A4F23-A068-4BAD-AC21-523A10A335AC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{92A7A7C4-CCDC-417E-84BA-B06F485630C5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{92A7A7C4-CCDC-417E-84BA-B06F485630C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{92A7A7C4-CCDC-417E-84BA-B06F485630C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{92A7A7C4-CCDC-417E-84BA-B06F485630C5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
