+++ app_dotnet/app_dotnet.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "app_dotnet", "app_dotnet\app_dotnet.csproj", "{134E5942-3DAE-4AA3-A4EB-B02FA38D4685}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "app_dotnet.tests", "app_dotnet.tests\app_dotnet.tests.csproj", "{4F163DF1-C5FC-423B-B7F7-61A80630C717}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{134E5942-3DAE-4AA3-A4EB-B02FA38D4685}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{134E5942-3DAE-4AA3-A4EB-B02FA38D4685}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{134E5942-3DAE-4AA3-A4EB-B02FA38D4685}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{134E5942-3DAE-4AA3-A4EB-B02FA38D4685}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4F163DF1-C5FC-423B-B7F7-61A80630C717}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4F163DF1-C5FC-423B-B7F7-61A80630C717}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4F163DF1-C5FC-423B-B7F7-61A80630C717}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4F163DF1-C5FC-423B-B7F7-61A80630C717}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
