+++ app_dotnet/app_dotnet.tests/UnitTest1.cs
@@ -1,6 +1,5 @@
 using Xunit;
-using App.Controllers;
-using app_cs.App.Services;
+using app_dotnet.Services;
 
 namespace Prime.UnitTests.Services;
 
@@ -12,6 +11,7 @@ public class DoItOrDont_ValidRequest_ShouldReturn200OK
     public void GetOption_ValidGet_ReturnsOneOfTwoOptions()
     {
         var response = _target.GetOption();
-        Assert.True(response == "Do it" || response == "Do not do it");
+        
+        Assert.True(response.YourOption is "Do it" or "Do not do it");
     }
 }
