+++ app_dotnet/app_dotnet/Program.cs
@@ -1,4 +1,4 @@
-using app_cs.App.Services;
+using app_dotnet.Services;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -20,8 +20,6 @@ if (app.Environment.IsDevelopment())
 }
 
 app.UseHttpsRedirection();
-
-app.UseAuthorization();
 
 app.MapControllers();
