+++ app_dotnet/app_dotnet/app_dotnet.http
@@ -1,0 +1,6 @@
+@app_dotnet_HostAddress = http://localhost:5228
+
+GET {{app_dotnet_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
